tao customer, update ảnh

+++ src/product/product.controller.ts
@@ -6,19 +6,30 @@ import {
   Get,
   Param,
   Post,
+  Put,
   Query,
+  UploadedFile,
   UploadedFiles,
   UseInterceptors,
 } from '@nestjs/common';
 import { Prop } from '@nestjs/mongoose';
-import { FileFieldsInterceptor } from '@nestjs/platform-express';
+import {
+  FileFieldsInterceptor,
+  FileInterceptor,
+} from '@nestjs/platform-express';
 import { CreateCategoryDto } from 'src/category/dto/create-category.dto';
 import { User } from 'src/user/model/user.schema';
 import { CreateProductDto } from './dto/create-product.dto';
 import { CloudinaryService } from 'src/cloudinary/cloudinary.service';
 import { ProductService } from './product.service';
-import { checkFileImage } from 'src/common/commom';
+import {
+  checkExtraFiles,
+  checkFileImage,
+  checkMainFile,
+} from 'src/common/commom';
 import { ParamPaginationDto } from 'src/common/param-pagination.dto';
+import { UpdateProductDto } from './dto/update-product.dto';
+import { Types } from 'mongoose';
 
 @Controller('products')
 export class ProducController {
@@ -41,29 +52,15 @@ export class ProducController {
   ) {
     checkFileImage(files);
 
-    if (files.main_image.length > 1) {
+    if (files.main_image && files.main_image.length > 1) {
       throw new BadRequestException('main_image chỉ nhận 1 file');
     }
-
     const newProduct = await this.productService.createProduct(product);
-
-    this.cloudinaryService
-      .uploadFile(files.main_image[0], 'products/' + newProduct._id)
-      .then((result) => {
-        this.productService.uploadMainImage(newProduct._id, {
-          image_id: result.public_id,
-          image_url: result.url,
-        });
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-
-    files.extra_images.forEach(async (file) => {
+    if (files.main_image) {
       this.cloudinaryService
-        .uploadFile(file, 'products/' + newProduct._id)
+        .uploadFile(files.main_image[0], 'products/' + newProduct._id)
         .then((result) => {
-          this.productService.uploadExtraImages(newProduct._id, {
+          this.productService.uploadMainImage(newProduct._id, {
             image_id: result.public_id,
             image_url: result.url,
           });
@@ -71,7 +68,23 @@ export class ProducController {
         .catch((error) => {
           console.log(error);
         });
-    });
+    }
+
+    if (files.extra_images) {
+      files.extra_images.forEach(async (file) => {
+        this.cloudinaryService
+          .uploadFile(file, 'products/' + newProduct._id)
+          .then((result) => {
+            this.productService.uploadExtraImages(newProduct._id, {
+              image_id: result.public_id,
+              image_url: result.url,
+            });
+          })
+          .catch((error) => {
+            console.log(error);
+          });
+      });
+    }
 
     return 'Đã tạo product thành công ';
   }
@@ -85,21 +98,78 @@ export class ProducController {
   async delete(@Param('id') id: string) {
     const product = await this.productService.deleteById(id);
 
-    this.cloudinaryService.deleteImage(product.image_id);
-    const deleteImages = async (imgs) => {
-      const deletePromises = imgs.map((image) =>
-        this.cloudinaryService.deleteImage(image.image_id),
-      );
-
-      try {
-        await Promise.all(deletePromises);
-      } catch (error) {
-        console.error('Error deleting images:', error);
-      }
-    };
-
-    await deleteImages(product.images);
+    await this.cloudinaryService.deleteById(`products/${product._id}`);
+    await this.cloudinaryService.deleteFolder(`products/${product._id}`);
 
     return 'Đã xóa product thành công';
+  }
+
+  @Put(':id')
+  update(@Param('id') id: string, @Body() product: UpdateProductDto) {
+    return this.productService.updateById(id, product);
+  }
+
+  @Put(':id/main_image')
+  @UseInterceptors(FileInterceptor('main_image'))
+  async updateImage(
+    @Param('id') id: string,
+    @UploadedFile() file: Express.Multer.File,
+  ) {
+    checkMainFile(file);
+
+    if (!file) {
+      throw new BadRequestException('Không nhận được file!');
+    }
+
+    const product = await this.productService.findById(id);
+
+    const result = await this.cloudinaryService.uploadFile(
+      file,
+      'products/' + product._id,
+    );
+
+    if (product.image_id) {
+      await this.cloudinaryService.deleteImage(product.image_id);
+    }
+    const newProduct = await this.productService.uploadMainImage(product._id, {
+      image_id: result.public_id,
+      image_url: result.url,
+    });
+
+    return newProduct;
+  }
+
+  @Get(':id')
+  getOne(@Param('id') id: string) {
+    return this.productService.findById(id);
+  }
+
+  // ham chuc nang aip xoa anh phu
+  @Put(':id/add_images')
+  @UseInterceptors(FileFieldsInterceptor([{ name: 'extra_images' }]))
+  async addImages(
+    @Param('id') id: string,
+    @UploadedFiles()
+    files: {
+      extra_images: Express.Multer.File[];
+    },
+  ) {
+    checkExtraFiles(files.extra_images);
+    if (!files.extra_images) {
+      throw new BadRequestException('Không nhận được file!');
+    } else {
+      files.extra_images.forEach((file) => {
+        this.cloudinaryService
+          .uploadFile(file, 'products/' + id)
+          .then((result) => {
+            this.productService.uploadExtraImages(new Types.ObjectId(id), {
+              image_id: result.public_id,
+              image_url: result.url,
+            });
+          });
+      });
+    }
+
+    return 'Đã ảnh phụ cho product này';
   }
 }
